# server/api-docs/openapi.yaml
swagger: "2.0"
info:
  description: "HTTP API for the Hilbert UI server."
  version: "1.0.0"
  title: "Hilbert UI"
  contact:
    email: "eric.londaits@imaginary.org"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api"
schemes:
- "http"
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /stations:
    get:
      tags:
        - "Stations"
      summary: "Get the list of stations"
      parameters:
        - in: query
          name: lastUpdateID
          type: integer
          description: "The last updateID received from calling this method. The server will only respond if or
            when the state of the stations changes (relative to the updateID). If this parameter is not sent the
            current state will be returned immediately. In case of a timeout before any updates an empty object
            will be returned"
      responses:
        200:
          description: Returns the list of stations with their current state.
          schema:
            type: object
            properties:
              updateID:
                type: number
                description: "A logical clock / incrementing id that can be sent back to the server to indicate
                  the last system state seen as part of the 'long poll' mechanism."
              stations:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: Station ID
                    name:
                      type: string
                      description: Station Name
                    description:
                      type: string
                      description: Station description
                    profile:
                      type: string
                      description: ID of the station profile the station belongs to
                    type:
                      type: string
                      description: Station type
                    state:
                      type: string
                      description: Current state of the station
                      enum: [unk, on, off, stopping, starting_station, starting_app, switching_app, error]
                    status:
                      type: string
                      description: "The station's status message which gives extra human readable
                        details about the state"
                    app:
                      type: string
                      description: ID of the app currently running in the station
                    default_app:
                      type: string
                      description: ID of the station's default app
                    compatible_apps:
                      type: array
                      description: IDs of the apps compatible with this station
                      items:
                        type: string
                    switching_app:
                      type: string
                      description: "ID of the app the station is currently switching to. An empty string
                        if the station is not currently changing apps."
                    icon:
                      type: string
                      description: URI of the icon of the currently running app.
  /stations/start:
    post:
      tags:
        - "Stations"
      summary: "Start a list of stations"
      consumes:
        - application/json
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
        400:
          description: Missing 'ids' argument
  /stations/stop:
    post:
      tags:
        - "Stations"
      summary: "Stop a list of stations"
      consumes:
        - application/json
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
        400:
          description: Missing 'ids' argument
  /stations/change_app:
    post:
      tags:
        - "Stations"
      summary: "Change the application in a list of stations"
      consumes:
        - application/json
      parameters:
        - in: body
          name: arguments
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
              app:
                description: App ID
                type: string
      responses:
        200:
          description: Success
        400:
          description: Missing argument
  /station/{id}/output:
    get:
      tags:
        - "Stations"
      summary: "Retrieve the station terminal output"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: ID of a station
      responses:
        200:
          description: Returns the terminal output
          schema:
            type: object
            properties:
              lines:
                type: array
                description: List of terminal output lines
                items:
                  type: string
...
  /test-backend/flags:
    get:
      tags:
        - "Test backend"
      summary: "Get the test backend flags"
      responses:
        200:
          description: Returns the values of the flags
          schema:
            type: object
            properties:
              testMode:
                type: boolean
              flags:
                $ref: "#/definitions/TestBackendFlags"

    post:
      tags:
        - "Test backend"
      summary: "Set the test backend flags"
      parameters:
        - in: body
          name: flags
          description: The flag values you want to set
          schema:
            $ref: "#/definitions/TestBackendFlags"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/TestBackendFlags"
  /test-backend/stations/unreachable:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations becoming unreachable
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
        400:
          description: Missing 'ids' argument
  /test-backend/stations/reachable:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations becoming reachable
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
        400:
          description: Missing 'ids' argument
  /test-backend/stations/stop-unexpectedly:
    post:
      tags:
        - "Test backend"
      summary: Simulate stations stopping unexpectedly
      parameters:
        - in: body
          name: arguments
          description: List of Station IDs
          schema:
            type: object
            properties:
              ids:
                description: List of station IDs
                type: array
                items:
                  type: string
      responses:
        200:
          description: Success
        400:
          description: Missing 'ids' argument
definitions:
...
  TestBackendFlags:
    type: object
    properties:
      sim-fail-cli:
        type: boolean
        default: false
      sim-unreachable:
        type: boolean
        default: false
      sim-timeout:
        type: boolean
        default: false